coverage: Collect HIR info during MIR building
The coverage instrumentor pass operates on MIR, but it needs access to some
source-related information that is normally not present in MIR.

Historically, that information was obtained by looking back at HIR during MIR
transformation. This patch arranges for the necessary information to be
collected ahead of time during MIR building instead.