Allow coercions from never-type when ref binding is involved
When we type-check a binding that uses ref patterns, we do not
perform coercions between the expression type and the pattern type.

However, this has the unfortunate result of disallow code like:
`let Foo { ref my_field } = diverging_expr();`. This code is accepted
without the `ref` keyword.

We now explicitly allow coercions when the expression type is !
In all other cases, we still avoid performing coersions.
This avoids causing broader changes in type-checking (with potential
soundness implications for 'ref mut'),