Detect `NulInCStr` error earlier.
By making it an `EscapeError` instead of a `LitError`. This makes it
more like the other errors produced during unescaping.

NOTE: this means these errors are issued earlier, before expansion,
which changes behaviour. The next commit will delay issue of this error
and others, reverting the behaviour change for this particular error.

One nice thing about this: the old approach had some code in
`report_lit_error` to calculate the span of the nul char from a range.
This code used a hardwired `+2` to account for the `c"` at the start of
a C string literal, but this should have changed to a `+3` for raw C
string literals to account for the `cr"`, which meant that the caret in
`cr"` nul error messages was one short of where it should have been. The
new approach doesn't need any of this and avoids the off-by-one error.