Move `report_lit_error` from `rustc_session` to `rustc_parse`.
It's a more logical spot for it, and will be a big help for the next
commit.

Doing this creates a new dependency from `rustc_ast_lowering` on
`rustc_parse`, but `rustc_ast_lowering` is clearly higher up the crate
graph, so this isn't a big deal.

One thing in favour of this change, is that two fluent labels were
duplicated across `rustc_session` and `rustc_parse`:
`invalid_literal_suffix` and `parse_not_supported`. This duplication is
now gone, so that's nice evidence that this is a reasonable change.