Emit `AssertLint` for overflow asserts found in MIR
This fixes the issue where no lints were firing for overflowing SHL and SHR ops.
That happened because `ConstPropagator::check_binary_op()`, which is responsible
for firing these lints, was skipping them because it couldn't find any `Shl()` or `Shr()`
ops in the main MIR. That in turn occurred because our const promotion pass
`PromoteTemps` removes these ops from the MIR when it runs.

With this change, the firing of these particular lints is predicated on the presence of
an assert terminator instead, which fortunately is not removed by `PromoteTemps`