Don't initialize stdout during io cleanup
Slightly simplify the `std::io::cleanup` code by using `OnceCell::get` instead
of checking initialization manually with `get_or_init`.

This also has the consequence of not initializing stdout with an empty buffer
if it hasn't been already, but I believe this is inconsequential given that
this code only runs at process exit.