Prevent `AtomicBool::fetch_and` and `AtomicBool::fetch_or` from CAS on x86 and x86_64
Currently, this operations generate Compare-And-Swap loop on x86 and x86_64.

Example:

```asm
    movzx        eax, byte ptr [rdi]
.LBB1_1:
    mov          ecx, eax
    and          cl, sil
    lock cmpxchg byte ptr [rdi], cl
    jne          .LBB1_1
    test         al, al
    setne        al
    ret
```

If we exploit the fact that AtomicBool can be either `0` and `1`, we can avoid this
loop, and, in cases when argument is known, any branching at all.

```asm
    test    esi, esi
    je      .LBB1_1
    mfence
    movzx   eax, byte ptr [rdi]
    jmp     .LBB1_2
.LBB1_1:
    xor     eax, eax
    xchg    byte ptr [rdi], al
.LBB1_2:
    test    al, al
    setne   al
    ret
```

We still need to use `swap` and `fetch_add` instead of `store` and `load` because we need to get old value
and apply provided memory ordering.

Related discussion: https://internals.rust-lang.org/t/should-we-change-fetch-logic-op-for-atomicbool-to-exploit-knowledge-about-it-contents/20098